feat: 成績計算邏輯完成（測試通過）AbstractGradeService.calculateGrade()

+++ SE-form-backend/src/main/java/com/isslab/se_form_backend/service/impl/GradeHelper.java
@@ -9,7 +9,7 @@ public class GradeHelper {
         return (gradeByScore - presenterGrade) / standardDeviation;
     }
 
-    public static int getGradeGap(Map<String, Integer> SCORE_MAP){
+    public static double getGradeGap(Map<String, Double> SCORE_MAP){
         return SCORE_MAP.get("A") - SCORE_MAP.get("C");
     }
